Refactor for smaller bundes on general pages'
