+++ pages/submit/upload.js
@@ -1,6 +1,6 @@
 import "localforage-getitems";
-import { NextSeo } from "next-seo";
 
+import { NextSeo } from "next-seo";
 import { connect } from "react-redux";
 import {
   loadTextContent,
@@ -9,13 +9,14 @@ import {
 import React, { useState } from "react";
 import localForage from "localforage";
 
-import { base64ToBlob, loadHeader } from "../../utils/storage";
+import { loadHeader } from "../../utils/storage/ls-composer";
 import ArticleSection from "../../core/components/pages/Article/components/ArticleSection";
 import ArticleWrapper from "../../core/components/pages/Article/components/ArticleWrapper";
 import HeaderLarge from "../../core/components/vignettes/HeaderLarge";
 import Link from "../../core/components/controls/Link";
 import Main from "../../core/components/layouts/Main";
 import SignIn from "../../user/components/pages/Account/SignIn";
+import base64ToBlob from "../../utils/base-64-to-blob";
 import isIncompleteDraft from "../../utils/editor/is-incomplete-draft";
 import uploadDraft from "../../utils/editor/upload-draft";
