+++ utils/base-64-to-blob.js
@@ -1,0 +1,19 @@
+// https://stackoverflow.com/questions/4998908/convert-data-uri-to-file-then-append-to-formdata
+export default string => {
+  if (string instanceof Blob) return string;
+  if (!string) return;
+
+  let byteString;
+  if (string.split(",")[0].indexOf("base64") >= 0)
+    byteString = atob(string.split(",")[1]);
+  else byteString = unescape(string.split(",")[1]);
+  const mimeString = string
+    .split(",")[0]
+    .split(":")[1]
+    .split(";")[0];
+  let ia = new Uint8Array(byteString.length);
+  for (let i = 0; i < byteString.length; i++) {
+    ia[i] = byteString.charCodeAt(i);
+  }
+  return new Blob([ia], { type: mimeString });
+};
